Introduce Component abstraction.
Currently we have a Program abstraction which encapsulates the required
functions of a runtime: Init, Update, and View. When creating a composed
program, it is not often useful to have an Init function. This
introduces a Component abstraction which defines Update and View, upon
which the Program interface inherits and adds Init.